Backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
azure-ai-documentintelligence
joblib
google-generativeai # For Gemini LLM
python-multipart